Initial (v0.1) commit of tracki3 script
Added tracki3 with rudimentary functionality and some commented
information pertaining to the script.

Requires python i3ipc, but otherwise works as intended.

Note: A lot could be added to make this script nicer, but I don't have
much motivation. It is a small script mean as a didactic tool for new
users.