Major changes to tracki3 (v0.2). Reapply
Added colors and switched to nicer glyphs for the tree.

Added commandline options to toggle colors, glyphs, and a few other
options that can be seen in -h.

Added in-built ability to have the tree update itself. You can set the
delay for the how often it will update via a commandline option.

General improvements to make the tree easier to understand and follow.

Tested all commandline parameters and the tracking.